+++ oxidauth-http/src/main.rs
@@ -6,11 +6,14 @@ pub mod server;
 use oxidauth_kernel::bootstrap::{BootstrapParams, BootstrapService};
 use oxidauth_usecases::bootstrap::SudoUserBootstrapUseCase;
 use std::{error::Error, sync::Arc};
+use tracing::info;
 
 use server::Server;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
+    info!("engaging oxidauth http server...");
+
     let subscriber = oxidauth_telemetry::get_subscriber(
         "oxidauth-http-api".into(),
         "INFO".into(),
@@ -29,10 +32,12 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync + 'static>> {
         .call(&BootstrapParams)
         .await?;
 
+    info!("starting server...");
     let addr = "0.0.0.0:80".parse()?;
 
     let server = Server::new(addr, provider);
 
+    info!("http booting...");
     server.start().await?;
 
     Ok(())
