+++ oxidauth-repository/src/totp_secrets/insert_totp_secret.rs
@@ -1,9 +1,12 @@
 pub use oxidauth_kernel::service::Service;
-use oxidauth_kernel::totp_secrets::create_totp_secret::{
-    CreateTotpSecretResponse, InsertTotpSecretParams,
-};
+use oxidauth_kernel::totp_secrets::create_totp_secret::CreateTotpSecretResponse;
 
 pub use crate::prelude::*;
+
+pub struct InsertTotpSecretParams {
+    pub user_id: Uuid,
+    pub secret_key: String,
+}
 
 pub trait InsertTotpSecretQuery:
     for<'a> Service<
